requirements.txt
boto3==1.7.24
botocore==1.10.24
docutils==0.14
jmespath==0.9.3
python-dateutil==2.7.3
s3transfer==0.1.13
six==1.11.0
tqdm==4.23.3
